# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pet Hotel & Grooming System API",
    "description": "Contract-first OpenAPI for the Pet Hotel & Grooming System.\n\n**S1 status:** endpoints are stubs (HTTP 501) with complete request/response schemas — this document is the single source of truth for backend + frontend. Business logic, state machines and real payment/notification land in S2+.\n\nAuthoritative data model: `diagrams/puml/class_diagram_plantuml.puml`. Internal modules (payment, notification, audit) are not exposed as endpoints.",
    "version": "0.1.0-S1"
  },
  "paths": {
...
    "/api/v1/pets": {
      "get": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "List my pets (FR-02.1)",
        "operationId": "list_pets_api_v1_pets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PetOut"
                  },
                  "type": "array",
                  "title": "Response List Pets Api V1 Pets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Create a pet profile (FR-02.1)",
        "operationId": "create_pet_api_v1_pets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pets/{pet_id}": {
      "get": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Get a pet profile",
        "operationId": "get_pet_api_v1_pets__pet_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Edit pet profile (FR-02.1/02.6)",
        "operationId": "update_pet_api_v1_pets__pet_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}/medical-records": {
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Add a medical record (FR-02.2)",
        "operationId": "add_medical_record_api_v1_pets__pet_id__medical_records_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalRecordCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicalRecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "List medical records (FR-02.2)",
        "operationId": "list_medical_records_api_v1_pets__pet_id__medical_records_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedicalRecordOut"
                  },
                  "title": "Response List Medical Records Api V1 Pets  Pet Id  Medical Records Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}/vaccinations": {
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Pre-enter a vaccination (FR-02.3)",
        "operationId": "add_vaccination_api_v1_pets__pet_id__vaccinations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VaccinationCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaccinationRecordOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "List vaccinations",
        "operationId": "list_vaccinations_api_v1_pets__pet_id__vaccinations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VaccinationRecordOut"
                  },
                  "title": "Response List Vaccinations Api V1 Pets  Pet Id  Vaccinations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}/vaccinations/{vaccination_id}/proof": {
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Upload vaccine proof document (FR-02.4)",
        "operationId": "upload_vaccine_proof_api_v1_pets__pet_id__vaccinations__vaccination_id__proof_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          },
          {
            "name": "vaccination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vaccination Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_vaccine_proof_api_v1_pets__pet_id__vaccinations__vaccination_id__proof_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VaccineProofDocumentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}/danger-level": {
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Mark danger level — Staff/Groomer: Low/Medium, Admin: High (FR-02.7)",
        "operationId": "mark_danger_level_api_v1_pets__pet_id__danger_level_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DangerLevelMarkIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pets/{pet_id}/unblock": {
      "post": {
        "tags": [
          "Pets & Health"
        ],
        "summary": "Unblock a pet — admin (FR-02.7)",
        "operationId": "unblock_pet_api_v1_pets__pet_id__unblock_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_vaccine_proof_api_v1_pets__pet_id__vaccinations__vaccination_id__proof_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_vaccine_proof_api_v1_pets__pet_id__vaccinations__vaccination_id__proof_post"
      },
...
      "DangerLevel": {
        "type": "string",
        "enum": [
          "None",
          "Low",
          "Medium",
          "High"
        ],
        "title": "DangerLevel"
      },
      "DangerLevelMarkIn": {
        "properties": {
          "dangerLevel": {
            "$ref": "#/components/schemas/DangerLevel",
            "examples": [
              "Medium"
            ]
          },
          "dangerNote": {
            "type": "string",
            "title": "Dangernote",
            "examples": [
              "曾對工作人員低吼"
            ]
          }
        },
        "type": "object",
        "required": [
          "dangerLevel",
          "dangerNote"
        ],
        "title": "DangerLevelMarkIn",
        "description": "FR-02.7 — Staff/Groomer: Low/Medium; Admin: High. Note required."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicalRecordCreateIn": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "examples": [
              "2025 年曾因皮膚過敏就診"
            ]
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "MedicalRecordCreateIn",
        "description": "FR-02.2 — append only; never modified/deleted afterwards."
      },
      "MedicalRecordOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "petId": {
            "type": "string",
            "format": "uuid",
            "title": "Petid"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "petId",
          "description",
          "createdAt"
        ],
        "title": "MedicalRecordOut"
      },
...
      "PetCreateIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "旺財"
            ]
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species",
            "examples": [
              "Dog"
            ]
          },
          "breed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed",
            "examples": [
              "柴犬"
            ]
          },
          "birthDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "chipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chipid",
            "examples": [
              "900123456789012"
            ]
          },
          "behaviorNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behaviornote",
            "examples": [
              "怕生，不喜歡被碰腳"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PetCreateIn",
        "description": "FR-02.1 / 02.6 — basic profile + optional behaviour note."
      },
      "PetOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid",
            "title": "Ownerid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed"
          },
          "birthDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "chipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chipid"
          },
          "behaviorNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behaviornote"
          },
          "dangerLevel": {
            "$ref": "#/components/schemas/DangerLevel"
          },
          "dangerNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dangernote"
          },
          "isBlocked": {
            "type": "boolean",
            "title": "Isblocked"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "name",
          "dangerLevel",
          "isBlocked",
          "createdAt"
        ],
        "title": "PetOut"
      },
      "PetUpdateIn": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species"
          },
          "breed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breed"
          },
          "birthDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "chipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chipid"
          },
          "behaviorNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Behaviornote"
          }
        },
        "type": "object",
        "title": "PetUpdateIn"
      },
...
      "VaccinationCreateIn": {
        "properties": {
          "vaccineName": {
            "type": "string",
            "title": "Vaccinename",
            "examples": [
              "狂犬病疫苗"
            ]
          },
          "administeredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Administeredat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat",
            "examples": [
              "2026-12-31"
            ]
          }
        },
        "type": "object",
        "required": [
          "vaccineName"
        ],
        "title": "VaccinationCreateIn",
        "description": "FR-02.3 — owner pre-enters vaccine + expiry (optional data)."
      },
      "VaccinationRecordOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "petId": {
            "type": "string",
            "format": "uuid",
            "title": "Petid"
          },
          "vaccineName": {
            "type": "string",
            "title": "Vaccinename"
          },
          "administeredAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Administeredat"
          },
          "expiresAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "status": {
            "$ref": "#/components/schemas/VaccinationStatus"
          },
          "verifiedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifiedat"
          },
          "verifiedById": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verifiedbyid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "petId",
          "vaccineName",
          "status"
        ],
        "title": "VaccinationRecordOut"
      },
      "VaccinationStatus": {
        "type": "string",
        "enum": [
          "Pending",
          "Valid",
          "Expired",
          "Rejected"
        ],
        "title": "VaccinationStatus"
      },
      "VaccineProofDocumentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "vaccinationRecordId": {
            "type": "string",
            "format": "uuid",
            "title": "Vaccinationrecordid"
          },
          "fileUrl": {
            "type": "string",
            "title": "Fileurl"
          },
          "uploadedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vaccinationRecordId",
          "fileUrl"
        ],
        "title": "VaccineProofDocumentOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Bearer JWT (enforced in S2)",
        "scheme": "bearer"
      }
    }
  },
...
}